backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend Financial Platform"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.124.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.20.0",
    "pre-commit>=4.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.9",
]
